Revert "Workshop detail"
