+++ src/Components/App/App.js
@@ -1,0 +1,42 @@
+import React, { Component } from 'react';
+import logo from './logo.svg';
+import './App.css';
+
+
+class App extends Component {
+  constructor(props) {
+    super(props)
+    this.state = {
+      isLoading: false,
+
+    }
+  }
+
+  componentDidMount() {
+  }
+
+  render() {
+    if (this.state.isLoading) return null
+    console.log("hello Homepage")
+    return (
+      <div className="App">
+        <header className="App-header">
+          <img src={logo} className="App-logo" alt="logo" />
+          <p>
+            Edit <code>src/App.js</code> and save to reload.
+            </p>
+          <a
+            className="App-link"
+            href="https://reactjs.org"
+            target="_blank"
+            rel="noopener noreferrer"
+          >
+            Learn React
+            </a>
+        </header>
+      </div>
+    )
+  }
+}
+
+export default App;
