--- src/Components/Workshop_detail/Button.css
@@ -1,7 +1,0 @@
-.button {
-    border-radius: 55px;
-    background-color: #F5A03C;
-    font-family: B612;
-    font-style: normal;
-    font-weight: normal;
-}
