--- src/Components/Workshop_detail/Instructor.css
@@ -1,22 +1,0 @@
-.instructor {
-    display: flex;
-    justify-content: left;
-    margin: 0;
-    padding-top: 2%;
-    padding-left: 1%;
-}
-
-.instructor-image {
-    border-radius: 100px;
-}
-
-.instructor-name {
-    font-family: B612;
-    font-style: normal;
-    font-weight: normal;
-}
-
-.instructor-detail {
-    margin: 2%;
-    align-self: center;
-}
