--- src/Components/Workshop_detail/Tag.css
@@ -1,8 +1,0 @@
-.Tag {
-    margin-left: 10%;
-}
-
-.link {
-    color: black;
-    text-decoration: underline;
-}
