--- src/Components/Workshop_detail/TagList.css
@@ -1,5 +1,0 @@
-.TagList {
-    display: flex;
-    justify-content: center;
-    margin: 2%;
-}
