--- src/Components/Workshop_detail/WorkshopDetailHeader.css
@@ -1,44 +1,0 @@
-.header {
-    display: flex;
-    justify-content: center;
-    margin: 0;
-    padding: 2% 10%;
-    background-color : #fafafa;
-}
-
-#box {
-    align-self: center;
-}
-
-.header-details {
-    width: 25em;
-    height: 25em;
-    display: flex;
-    justify-content: center;
-    flex-direction: column;
-    background-color: rgba(204, 115, 10, 0.12);
-    padding: 2%;
-}
-.header-detail {
-    align-self: left;
-    margin: 0.5em;
-    font-family: B612;
-    font-style: normal;
-    font-weight: normal;
-}
-
-.body {
-    height: auto;
-    display: flex;
-    margin: 0;
-    background-color: #FFE8C7;
-}
-
-.body-details {
-    width: 100%;
-    height: 100%;
-    display: flex;
-    align-self: center;
-    justify-content: center;
-    margin: auto;
-}
